chore: restructure benchmarks
