+++ benchmarks/cookies/Is-ctl-excluding-htab.mjs
@@ -1,5 +1,5 @@
 import { bench, group, run } from 'mitata'
-import { isCTLExcludingHtab } from '../lib/web/cookies/util.js'
+import { isCTLExcludingHtab } from '../../lib/web/cookies/util.js'
 
 const valid = 'Space=Cat; Secure; HttpOnly; Max-Age=2'
 const invalid = 'Space=Cat; Secure; HttpOnly; Max-Age=2\x7F'
